refactor code a little bit
